Add useLocale to US to read Double-values in Matlab-Output correctly
The Scanner-class of some non-US Java-versions don´t interpret floating
point numbers like 3.52e-01 correctly. Because in some languages is a
comma instead of a point used, i. e. 3,52e-01. So if you take the
Scanner-class for reading such numbers(Matlab-output in this case) you
must localize the object to the US.